refactor: move 3 Smithery provider tests from CLI to SDK (Phase 3.2)
Move provider-smithery, smithery-auth, and smithery-publish tests to
packages/sdk/test/providers/. Update imports and error codes:
- E_UNSUPPORTED -> E_PROVIDER_UNAVAILABLE
- E_AUTH_REQUIRED -> E_PROVIDER_UNAVAILABLE
- E_AUTH_FAILED -> E_PROVIDER_UNAVAILABLE
- E_MISSING_NAME -> E_VALIDATION_FAILED
- E_MISSING_SOURCE -> E_VALIDATION_FAILED
- E_SERVER_NOT_FOUND -> E_COMPONENT_NOT_FOUND
- E_RATE_LIMITED -> E_PROVIDER_TIMEOUT
- E_API_ERROR -> E_PROVIDER_UNAVAILABLE
- E_INVALID_NAME -> E_VALIDATION_FAILED
- E_MISSING_MANIFEST, E_INVALID_MANIFEST -> E_VALIDATION_FAILED
- E_DEPLOY_TIMEOUT -> E_PROVIDER_TIMEOUT
- E_NETWORK -> E_PROVIDER_UNAVAILABLE
- E_TIMEOUT -> E_PROVIDER_TIMEOUT
- hint -> detail (SdkError field rename)
